requirements.txt
flask>=3.0
authlib>=1.3
requests
pycups
python-dotenv
gunicorn
flask-cors
imapclient
pyyaml
python-magic
werkzeug>=3.0
